chatBotMCP/requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
mcp
numpy
fastmcp
torch
transformers
accelerate
groq
sentence-transformers